# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BeCreative student API
  version: "1.0.0"
  description: |
    Versioned JSON API for the student iOS and Android apps. Teachers stay on the web.
    Web server actions and these routes call the same service layer.
    Authenticated routes expect `Authorization: Bearer bc_<token>`.
    Tokens last 30 days and are stored as a SHA-256 hash.
    Send `X-Platform: ios` or `X-Platform: android` (or `?platform=`). The default is `web`.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
...
paths:
...
  /bookings:
    get:
      summary: My bookings
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Booking list
    post:
      summary: Book a session
      description: Same bookSession service as the web. paymentSheet true returns a Stripe PaymentIntent client secret for PaymentSheet. Otherwise a Checkout Session URL when Stripe is on. Pay-at-studio when Stripe is off.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId: { type: string }
                classId: { type: string }
                series: { type: boolean }
                code: { type: string }
                payWith: { type: string, description: "cash, pack:<id>, or membership:<id>" }
                policyAccepted: { type: boolean }
                paymentSheet: { type: boolean }
      responses:
        "200":
          description: Order, clientSecret, or checkoutUrl
        "400":
          description: Booking error
...
  /notifications:
    get:
      summary: In-app notifications
      security: [{ bearer: [] }]
      responses:
        "200":
          description: List
...